Refactored channels to allow for subclassing
